% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visvow.R
\name{normalizeDuration}
\alias{normalizeDuration}
\title{Normalize duration}
\usage{
normalizeDuration(vowelTab, replyNormal)
}
\arguments{
\item{vowelTab}{A data frame containing acoustic vowel measurements; the format is described at https://www.visiblevowels.org/#help .}

\item{replyNormal}{Choose from: "none", "Lobanov".}
}
\description{
Normalize duration of vowels.
}
