% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visvow.R
\name{normalizeFormants}
\alias{normalizeFormants}
\title{Normalize vowel formants.}
\usage{
normalizeFormants(vowelTab, replyScale, replyNormal, replyTimesN)
}
\arguments{
\item{vowelTab}{A data frame containing acoustic vowel measurements; the format is described at https://www.visiblevowels.org/#help .}

\item{replyScale}{Choose from: "Hz", "bark I", "bark II", "bark III", "ERB I", "ERB II", "ERB III", "ln", "mel I", "mel II", "ST".}

\item{replyNormal}{Choose from: "none", "Peterson", "Sussman", "Syrdal & Gopal", "Miller", "Thomas & Kendall", "Gerstman", "Lobanov", "Watt & Fabricius", "Fabricius et al.", "Bigham", "Heeringa & Van de Velde I", "Heeringa & Van de Velde II", "Nearey I", "Nearey II", "Barreda & Nearey I", "Barreda & Nearey II", "Labov log-mean I", "Labov log-geomean I", "Labov log-mean II", "Labov log-geomean II", "Johnson".}

\item{replyTimesN}{If measurements are provided for multiple time points per vowel, provide the indices of the time points that should be included when descriptives such as minimum, maximum, mean and standard deviation are calculated by some normalization methods; when there is just one time point, give index 1; a set of multiple indices are given as a vector, for example, when there are three indices and you want the first and third index be used, give c(1,3).}
}
\description{
Scale and/or normalize formants F1, F2 and F3.
}
