% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visvow.R
\name{visvow}
\alias{visvow}
\title{Visible Vowels}
\usage{
visvow()
}
\description{
Visible Vowels is an app that visualizes vowel variation in f0, F1, F2, F3 and duration.
}
\details{
\code{visvow()} opens Visible Vowels in your default web browser.
}
\examples{
visvow

}
\references{
\insertRef{visvow:2017}{visvow}
}
\seealso{
The Help tab in the app provides more information about the format of the input file.
}
